+++ src/components/navigation/Breadcrumbs/index.tsx
@@ -1,7 +1,6 @@
 import { useMediaQuery } from "@hooks/useMediaQuery";
 
-import { BreadcrumbLink } from "@navigation/BreadcrumbLink";
-
+import { BreadcrumbLink } from "./BreadcrumbLink";
 import { BreadcrumbEllipsis } from "./BreadcrumbEllipsis";
 import { Sizes } from "./props";
 import { StyledBreadcrumbs } from "./styles";
